Remove functions related to v_delta in LCAO_Deepks; Remove some redundent variables.
