+++ source/module_hamilt_lcao/module_deepks/CMakeLists.txt
@@ -3,20 +3,21 @@ if(ENABLE_DEEPKS)
       LCAO_deepks.cpp
       deepks_force.cpp
       deepks_orbital.cpp
+      deepks_orbpre.cpp
+      deepks_vdpre.cpp
+      deepks_hmat.cpp
       LCAO_deepks_io.cpp
       LCAO_deepks_mpi.cpp
       LCAO_deepks_pdm.cpp
       LCAO_deepks_phialpha.cpp
       LCAO_deepks_torch.cpp
       LCAO_deepks_vdelta.cpp
-      deepks_hmat.cpp
       LCAO_deepks_interface.cpp
-      deepks_orbpre.cpp
       cal_gdmx.cpp
       cal_gedm.cpp
       cal_gvx.cpp      
       cal_descriptor.cpp
-      v_delta_precalc.cpp
+      
   )
 
   add_library(
