+++ source/module_hamilt_lcao/module_deepks/LCAO_deepks.cpp
@@ -343,107 +343,6 @@ void LCAO_Deepks::allocate_V_delta(const int nat, const int nks)
     return;
 }
 
-void LCAO_Deepks::init_v_delta_pdm_shell(const int nks, const int nlocal)
-{
-    const int mn_size = (2 * this->lmaxd + 1) * (2 * this->lmaxd + 1);
-    if (nks == 1)
-    {
-        this->v_delta_pdm_shell = new double****[nks];
-        for (int iks = 0; iks < nks; iks++)
-        {
-            this->v_delta_pdm_shell[iks] = new double***[nlocal];
-
-            for (int mu = 0; mu < nlocal; mu++)
-            {
-                this->v_delta_pdm_shell[iks][mu] = new double**[nlocal];
-
-                for (int nu = 0; nu < nlocal; nu++)
-                {
-                    this->v_delta_pdm_shell[iks][mu][nu] = new double*[this->inlmax];
-
-                    for (int inl = 0; inl < this->inlmax; inl++)
-                    {
-                        this->v_delta_pdm_shell[iks][mu][nu][inl] = new double[mn_size];
-                        ModuleBase::GlobalFunc::ZEROS(v_delta_pdm_shell[iks][mu][nu][inl], mn_size);
-                    }
-                }
-            }
-        }
-    }
-    else
-    {
-        this->v_delta_pdm_shell_complex = new std::complex<double>****[nks];
-        for (int iks = 0; iks < nks; iks++)
-        {
-            this->v_delta_pdm_shell_complex[iks] = new std::complex<double>***[nlocal];
-
-            for (int mu = 0; mu < nlocal; mu++)
-            {
-                this->v_delta_pdm_shell_complex[iks][mu] = new std::complex<double>**[nlocal];
-
-                for (int nu = 0; nu < nlocal; nu++)
-                {
-                    this->v_delta_pdm_shell_complex[iks][mu][nu] = new std::complex<double>*[this->inlmax];
-
-                    for (int inl = 0; inl < this->inlmax; inl++)
-                    {
-                        this->v_delta_pdm_shell_complex[iks][mu][nu][inl] = new std::complex<double>[mn_size];
-                        ModuleBase::GlobalFunc::ZEROS(v_delta_pdm_shell_complex[iks][mu][nu][inl], mn_size);
-                    }
-                }
-            }
-        }
-    }
-
-    return;
-}
-
-void LCAO_Deepks::del_v_delta_pdm_shell(const int nks, const int nlocal)
-{
-    if (nks == 1)
-    {
-        for (int iks = 0; iks < nks; iks++)
-        {
-            for (int mu = 0; mu < nlocal; mu++)
-            {
-                for (int nu = 0; nu < nlocal; nu++)
-                {
-                    for (int inl = 0; inl < this->inlmax; inl++)
-                    {
-                        delete[] this->v_delta_pdm_shell[iks][mu][nu][inl];
-                    }
-                    delete[] this->v_delta_pdm_shell[iks][mu][nu];
-                }
-                delete[] this->v_delta_pdm_shell[iks][mu];
-            }
-            delete[] this->v_delta_pdm_shell[iks];
-        }
-        delete[] this->v_delta_pdm_shell;
-    }
-    else
-    {
-        for (int iks = 0; iks < nks; iks++)
-        {
-            for (int mu = 0; mu < nlocal; mu++)
-            {
-                for (int nu = 0; nu < nlocal; nu++)
-                {
-                    for (int inl = 0; inl < this->inlmax; inl++)
-                    {
-                        delete[] this->v_delta_pdm_shell_complex[iks][mu][nu][inl];
-                    }
-                    delete[] this->v_delta_pdm_shell_complex[iks][mu][nu];
-                }
-                delete[] this->v_delta_pdm_shell_complex[iks][mu];
-            }
-            delete[] this->v_delta_pdm_shell_complex[iks];
-        }
-        delete[] this->v_delta_pdm_shell_complex;
-    }
-
-    return;
-}
-
 template <typename TK>
 void LCAO_Deepks::dpks_cal_e_delta_band(const std::vector<std::vector<TK>>& dm, const int nks)
 {
