+++ source/module_hamilt_lcao/module_deepks/LCAO_deepks_torch.cpp
@@ -2,7 +2,7 @@
 // including loading of model and calculating gradients
 // as well as subroutines that prints the results for checking
 
-// The file contains 8 subroutines:
+// The file contains 3 subroutines:
 //  cal_gvepsl : gvepsl is used for training with stress label, which is derivative of
 //        descriptors wrt strain tensor, calculated by
 //        d(des)/d\epsilon_{ab} = d(pdm)/d\epsilon_{ab} * d(des)/d(pdm) = gdm_epsl * gvdm
@@ -10,8 +10,6 @@
 //  cal_gevdm : d(des)/d(pdm)
 //        calculated using torch::autograd::grad
 //  load_model : loads model for applying V_delta
-//  prepare_phialpha : prepare phialpha for outputting npy file
-//  prepare_gevdm : prepare gevdm for outputting npy file
 
 #ifdef __DEEPKS
 
@@ -155,290 +153,5 @@ void LCAO_Deepks::load_model(const std::string& deepks_model)
     }
     return;
 }
-
-// prepare_phialpha and prepare_gevdm for deepks_v_delta = 2
-template <typename TK>
-void LCAO_Deepks::prepare_phialpha(const int nlocal,
-                                   const int nat,
-                                   const int nks,
-                                   const std::vector<ModuleBase::Vector3<double>>& kvec_d,
-                                   const UnitCell& ucell,
-                                   const LCAO_Orbitals& orb,
-                                   const Grid_Driver& GridD)
-{
-    ModuleBase::TITLE("LCAO_Deepks", "prepare_phialpha");
-    int nlmax = this->inlmax / nat;
-    int mmax = 2 * this->lmaxd + 1;
-    if constexpr (std::is_same<TK, double>::value)
-    {
-        this->phialpha_tensor = torch::zeros({nat, nlmax, nks, nlocal, mmax}, torch::kFloat64); // support gamma-only
-    }
-    else
-    {
-        this->phialpha_tensor = torch::zeros({nat, nlmax, nks, nlocal, mmax}, torch::kComplexDouble); // support multi-k
-    }
-
-    // cutoff for alpha is same for all types of atoms
-    const double Rcut_Alpha = orb.Alpha[0].getRcut();
-
-    for (int T0 = 0; T0 < ucell.ntype; T0++)
-    {
-        Atom* atom0 = &ucell.atoms[T0];
-        for (int I0 = 0; I0 < atom0->na; I0++)
-        {
-            // iat: atom index on which |alpha> is located
-            const int iat = ucell.itia2iat(T0, I0);
-            const ModuleBase::Vector3<double> tau0 = atom0->tau[I0];
-            GridD.Find_atom(ucell, atom0->tau[I0], T0, I0);
-
-            // outermost loop : find all adjacent atoms
-            for (int ad = 0; ad < GridD.getAdjacentNum() + 1; ++ad)
-            {
-                const int T1 = GridD.getType(ad);
-                const int I1 = GridD.getNatom(ad);
-                const int ibt = ucell.itia2iat(T1, I1);
-                const int start1 = ucell.itiaiw2iwt(T1, I1, 0);
-                const double Rcut_AO1 = orb.Phi[T1].getRcut();
-
-                const ModuleBase::Vector3<double> tau1 = GridD.getAdjacentTau(ad);
-                const Atom* atom1 = &ucell.atoms[T1];
-                const int nw1_tot = atom1->nw * PARAM.globalv.npol;
-
-                const double dist1 = (tau1 - tau0).norm() * ucell.lat0;
-
-                if (dist1 > Rcut_Alpha + Rcut_AO1)
-                {
-                    continue;
-                }
-
-                ModuleBase::Vector3<int> dR(GridD.getBox(ad).x, GridD.getBox(ad).y, GridD.getBox(ad).z);
-
-                if constexpr (std::is_same<TK, std::complex<double>>::value)
-                {
-                    if (this->phialpha[0]->find_matrix(iat, ibt, dR.x, dR.y, dR.z) == nullptr)
-                    {
-                        continue;
-                    }
-                }
-
-                // middle loop : all atomic basis on the adjacent atom ad
-                for (int iw1 = 0; iw1 < nw1_tot; ++iw1)
-                {
-                    const int iw1_all = start1 + iw1;
-                    const int iw1_local = pv->global2local_row(iw1_all);
-                    const int iw2_local = pv->global2local_col(iw1_all);
-                    if (iw1_local < 0 || iw2_local < 0)
-                    {
-                        continue;
-                    }
-                    hamilt::BaseMatrix<double>* overlap = phialpha[0]->find_matrix(iat, ibt, dR);
-
-                    for (int ik = 0; ik < nks; ik++)
-                    {
-                        std::complex<double> kphase = std::complex<double>(1.0, 0.0);
-                        if constexpr (std::is_same<TK, std::complex<double>>::value)
-                        {
-                            const double arg = -(kvec_d[ik] * ModuleBase::Vector3<double>(dR)) * ModuleBase::TWO_PI;
-                            double sinp, cosp;
-                            ModuleBase::libm::sincos(arg, &sinp, &cosp);
-                            kphase = std::complex<double>(cosp, sinp);
-                        }
-                        int ib = 0;
-                        int nl = 0;
-                        for (int L0 = 0; L0 <= orb.Alpha[0].getLmax(); ++L0)
-                        {
-                            for (int N0 = 0; N0 < orb.Alpha[0].getNchi(L0); ++N0)
-                            {
-                                const int nm = 2 * L0 + 1;
-
-                                for (int m1 = 0; m1 < nm; ++m1) // nm = 1 for s, 3 for p, 5 for d
-                                {
-                                    if constexpr (std::is_same<TK, double>::value)
-                                    {
-                                        this->phialpha_tensor[iat][nl][ik][iw1_all][m1]
-                                            = overlap->get_value(iw1, ib + m1);
-                                    }
-                                    else
-                                    {
-                                        std::complex<double> nlm_phase = overlap->get_value(iw1, ib + m1) * kphase;
-                                        torch::Tensor nlm_tensor
-                                            = torch::tensor({nlm_phase.real(), nlm_phase.imag()}, torch::kDouble);
-                                        torch::Tensor complex_tensor = torch::complex(nlm_tensor[0], nlm_tensor[1]);
-                                        this->phialpha_tensor[iat][nl][ik][iw1_all][m1] = complex_tensor;
-                                    }
-                                }
-                                ib += nm;
-                                nl++;
-                            }
-                        }
-                    } // end ik
-                }     // end iw
-            }         // end ad
-        }             // end I0
-    }                 // end T0
-
-#ifdef __MPI
-    TK msg[mmax];
-    for (int iat = 0; iat < nat; iat++)
-    {
-        for (int nl = 0; nl < nlmax; nl++)
-        {
-            for (int ik = 0; ik < nks; ik++)
-            {
-                for (int mu = 0; mu < nlocal; mu++)
-                {
-                    for (int m = 0; m < mmax; m++)
-                    {
-                        if constexpr (std::is_same<TK, double>::value)
-                        {
-                            msg[m] = this->phialpha_tensor[iat][nl][ik][mu][m].item().toDouble();
-                        }
-                        else
-                        {
-                            auto tensor_value = this->phialpha_tensor.index({iat, nl, ik, mu, m});
-                            msg[m] = std::complex<double>(torch::real(tensor_value).item<double>(),
-                                                          torch::imag(tensor_value).item<double>());
-                        }
-                    }
-                    Parallel_Reduce::reduce_all(msg, mmax);
-                    for (int m = 0; m < mmax; m++)
-                    {
-                        if constexpr (std::is_same<TK, double>::value)
-                        {
-                            this->phialpha_tensor[iat][nl][ik][mu][m] = msg[m];
-                        }
-                        else
-                        {
-                            torch::Tensor msg_tensor = torch::tensor({msg[m].real(), msg[m].imag()}, torch::kDouble);
-                            torch::Tensor complex_tensor = torch::complex(msg_tensor[0], msg_tensor[1]);
-                            this->phialpha_tensor[iat][nl][ik][mu][m] = complex_tensor;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-#endif
-}
-
-template <typename TK>
-void LCAO_Deepks::check_vdp_phialpha(const int nat, const int nks, const int nlocal)
-{
-    std::ofstream ofs("vdp_phialpha.dat");
-    ofs << std::setprecision(10);
-
-    int nlmax = this->inlmax / nat;
-    int mmax = 2 * this->lmaxd + 1;
-    for (int iat = 0; iat < nat; iat++)
-    {
-        for (int nl = 0; nl < nlmax; nl++)
-        {
-            for (int iks = 0; iks < nks; iks++)
-            {
-                for (int mu = 0; mu < nlocal; mu++)
-                {
-                    for (int m = 0; m < mmax; m++)
-                    {
-                        if constexpr (std::is_same<TK, double>::value)
-                        {
-                            ofs << this->phialpha_tensor.index({iat, nl, iks, mu, m}).item().toDouble() << " ";
-                        }
-                        else
-                        {
-                            auto tensor_value = this->phialpha_tensor.index({iat, nl, iks, mu, m});
-                            ofs << std::complex<double>(torch::real(tensor_value).item<double>(),
-                                                        torch::imag(tensor_value).item<double>())
-                                << " ";
-                        }
-                    }
-                }
-            }
-            ofs << std::endl;
-        }
-    }
-    ofs.close();
-}
-
-void LCAO_Deepks::prepare_gevdm(const int nat, const LCAO_Orbitals& orb)
-{
-    int nlmax = this->inlmax / nat;
-    int mmax = 2 * this->lmaxd + 1;
-    this->gevdm_tensor = torch::zeros({nat, nlmax, mmax, mmax, mmax}, torch::TensorOptions().dtype(torch::kFloat64));
-
-    std::vector<torch::Tensor> gevdm;
-    this->cal_gevdm(nat, gevdm);
-
-    int nl = 0;
-    for (int L0 = 0; L0 <= orb.Alpha[0].getLmax(); ++L0)
-    {
-        for (int N0 = 0; N0 < orb.Alpha[0].getNchi(L0); ++N0)
-        {
-            for (int iat = 0; iat < nat; iat++)
-            {
-                const int nm = 2 * L0 + 1;
-                for (int v = 0; v < nm; ++v) // nm = 1 for s, 3 for p, 5 for d
-                {
-                    for (int m = 0; m < nm; ++m)
-                    {
-                        for (int n = 0; n < nm; ++n)
-                        {
-                            this->gevdm_tensor[iat][nl][v][m][n] = gevdm[nl][iat][v][m][n];
-                        }
-                    }
-                }
-            }
-            nl++;
-        }
-    }
-    assert(nl == nlmax);
-}
-
-void LCAO_Deepks::check_vdp_gevdm(const int nat)
-{
-    std::ofstream ofs("vdp_gevdm.dat");
-    ofs << std::setprecision(10);
-
-    int nlmax = this->inlmax / nat;
-    int mmax = 2 * this->lmaxd + 1;
-    for (int iat = 0; iat < nat; iat++)
-    {
-        for (int nl = 0; nl < nlmax; nl++)
-        {
-            for (int v = 0; v < mmax; v++)
-            {
-                for (int m = 0; m < mmax; m++)
-                {
-                    for (int n = 0; n < mmax; n++)
-                    {
-                        ofs << this->gevdm_tensor.index({iat, nl, v, m, n}).item().toDouble() << " ";
-                    }
-                }
-            }
-            ofs << std::endl;
-        }
-    }
-    ofs.close();
-}
-
-template void LCAO_Deepks::prepare_phialpha<double>(const int nlocal,
-                                                    const int nat,
-                                                    const int nks,
-                                                    const std::vector<ModuleBase::Vector3<double>>& kvec_d,
-                                                    const UnitCell& ucell,
-                                                    const LCAO_Orbitals& orb,
-                                                    const Grid_Driver& GridD);
-
-template void LCAO_Deepks::prepare_phialpha<std::complex<double>>(
-    const int nlocal,
-    const int nat,
-    const int nks,
-    const std::vector<ModuleBase::Vector3<double>>& kvec_d,
-    const UnitCell& ucell,
-    const LCAO_Orbitals& orb,
-    const Grid_Driver& GridD);
-
-template void LCAO_Deepks::check_vdp_phialpha<double>(const int nat, const int nks, const int nlocal);
-template void LCAO_Deepks::check_vdp_phialpha<std::complex<double>>(const int nat, const int nks, const int nlocal);
 
 #endif
